+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ## [1.2.8] - 2025-07-29
+### Fixed
+- Updated CI
+- Fixed some SWIG warnings
 
 ## [1.2.7] - 2025-07-18
 ### Fixed